refactor(team): extract REST routes into lib/team-routes
Completes the team/index split (WS6). The ~970-line `populateTeamRoutes` —
every team REST route — moves into lib/team-routes.ts.

The handlers were already ctx-parameterized (they take PluginContextLite per
request), and the helper groups they leaned on (runtime-agents, team-settings,
agent-status) are now importable libs from the prior PRs. That left exactly one
piece of module-scope coupling: `indexAgentStatic` (the fire-and-forget search
upsert, which is bound to the live plugin context). It's injected as a single
`deps` parameter:

  populateTeamRoutes(teamRoutes, { indexAgentStatic })

`pluginCtx` and `batchIndexAgents` turned out to be activate-only, not used by
any route, so nothing else had to be threaded through.

index.ts: 1,685 -> 693 lines (2,312 -> 693 across the three team PRs, ~70%).
What remains in index.ts is activate()/onReady/onShutdown, the search-indexing
wiring (pluginCtx/indexAgentStatic/agentToSearchDocStatic/batchIndexAgents),
exec tools, and health-check registration.

Verified: typecheck + lint clean; team suite 183/0; full suite 5123/0; binary
build; isolated-home boot smoke (Team activates; GET /api/plugins/team/,
/teams, and /settings all 200 through the relocated handlers).
